CIRC-367: Code refactoring
