+++ src/main/java/org/folio/circulation/domain/anonymization/service/AnonymizationCheckersService.java
@@ -2,28 +2,34 @@ package org.folio.circulation.domain.anonymization.service;
 
 import static org.folio.circulation.domain.anonymization.LoanAnonymizationRecords.CAN_BE_ANONYMIZED_KEY;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 import org.apache.commons.collections4.multimap.HashSetValuedHashMap;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.anonymization.checkers.AnonymizationChecker;
+import org.folio.circulation.domain.anonymization.checkers.AnonymizeLoansImmediatelyChecker;
+import org.folio.circulation.domain.anonymization.checkers.AnonymizeLoansWithFeeFinesImmediatelyChecker;
+import org.folio.circulation.domain.anonymization.checkers.FeesAndFinesClosePeriodChecker;
+import org.folio.circulation.domain.anonymization.checkers.LoanClosePeriodChecker;
+import org.folio.circulation.domain.anonymization.checkers.NeverAnonymizeLoansChecker;
+import org.folio.circulation.domain.anonymization.checkers.NeverAnonymizeLoansWithFeeFinesChecker;
+import org.folio.circulation.domain.anonymization.checkers.NoAssociatedFeesAndFinesChecker;
 import org.folio.circulation.domain.anonymization.config.LoanAnonymizationConfigurationForTenant;
 
 public class AnonymizationCheckersService {
 
   private final LoanAnonymizationConfigurationForTenant config;
-  private AnonymizationCheckersProvider checkersProvider;
 
   public AnonymizationCheckersService(
       LoanAnonymizationConfigurationForTenant config) {
     this.config = config;
-    checkersProvider = new AnonymizationCheckersProvider(config);
   }
 
   public AnonymizationCheckersService() {
     this(null);
-    checkersProvider = new AnonymizationCheckersProvider();
   }
 
   public HashSetValuedHashMap<String, String> segregateLoans(Collection<Loan> loans) {
@@ -46,12 +52,66 @@ public class AnonymizationCheckersService {
 
   private List<AnonymizationChecker> getAnonymizationCheckers(Loan loan) {
     if (config == null) {
-      return checkersProvider.getGeneralCheckers();
+      return getDefaultCheckers();
     }
-    if (!loan.getAccounts().isEmpty() && config.treatLoansWithFeesAndFinesDifferently()) {
-      return checkersProvider.getFeesAndFinesCheckers();
+    if (loan.hasAssociatedFeesAndFines() && config.treatLoansWithFeesAndFinesDifferently()) {
+      return getFeesAndFinesCheckersFromLoanHistory();
     } else {
-      return checkersProvider.getClosedLoansCheckers();
+      return getClosedLoansCheckersFromLoanHistory();
     }
+  }
+
+  private List<AnonymizationChecker> getDefaultCheckers() {
+    return Collections.singletonList(new NoAssociatedFeesAndFinesChecker());
+  }
+
+  private List<AnonymizationChecker> getClosedLoansCheckersFromLoanHistory() {
+    List<AnonymizationChecker> result = new ArrayList<>();
+    if (config == null) {
+      return result;
+    }
+
+    switch (config.getLoanClosingType()) {
+    case IMMEDIATELY:
+      result.add(new AnonymizeLoansImmediatelyChecker());
+      break;
+    case INTERVAL:
+      result.add(new LoanClosePeriodChecker(
+          config.getLoanClosePeriod()));
+      break;
+    case UNKNOWN:
+    case NEVER:
+      result.add(new NeverAnonymizeLoansChecker());
+      break;
+    default:
+      return result;
+    }
+
+    return result;
+  }
+
+  private List<AnonymizationChecker> getFeesAndFinesCheckersFromLoanHistory() {
+    List<AnonymizationChecker> result = new ArrayList<>();
+    if (config == null || !config
+        .treatLoansWithFeesAndFinesDifferently()) {
+      return result;
+    }
+    switch (config.getFeesAndFinesClosingType()) {
+    case IMMEDIATELY:
+      result.add(new AnonymizeLoansWithFeeFinesImmediatelyChecker());
+      break;
+    case INTERVAL:
+      result.add(new FeesAndFinesClosePeriodChecker(
+          config.getFeeFineClosePeriod()));
+      break;
+    case UNKNOWN:
+    case NEVER:
+      result.add(new NeverAnonymizeLoansWithFeeFinesChecker());
+      break;
+    default:
+      return result;
+    }
+
+    return result;
   }
 }
